corrected redirecting logic for new users, for when you register to a new plan and restricted access to some pages depending on whether the user has a plan or not
