+++ src/components/protectedRoutes/plans/Plans.js
@@ -4,6 +4,8 @@ import monthlyPlanImage from '../../../assets/image02.jpg';
 import PlanDetails from './PlanDetails';
 import { useContext } from 'react';
 import UserContext from '../../../contexts/UserContext';
+import { useNavigate } from 'react-router';
+import { useEffect } from 'react/cjs/react.development';
 
 const plans = [
     {
@@ -22,6 +24,14 @@ const plans = [
 
 export default function Plans() {
     const { user } = useContext(UserContext);
+    const navigate = useNavigate();
+
+    useEffect(() => {
+        if (localStorage.getItem('plano') === 'true') {
+            navigate('/detalhes-plano');
+        }
+    }, []);
+
     return (
         <PageContainer>
             <Title>Bom te ver por aqui, {user.name}</Title>
